Turned off mesh renderer for the sphere
how did i turn it off? apparently i was too clingy and caring for it, the mesh rendered wanted some space and i couldnt give that to it. Idk man i tried my best to be everything for him but I guess i was too much, things will never be the same between, I think this is the end of me and mesh rendered </3 i loved it with all my heart and yet I wasnt enough - i was too much actually but you get the point. I guess this is goodbye mesh rendered, we had a good run but this is the end.